requirements.txt
torch
pandas
scikit-learn
numpy
beautifulsoup4
python-dotenv
fastapi
uvicorn
requests
joblib